Use correct variable names for expansion and rarity.
So we're clear: if a card has appearned in multiple expansions,
all_sets will contain all of these.  If it has only appeared in one,
then expansion and rarity will be populated separately and set_rarity
will combine these in the correct format.  When showing a card,
display_set will pick what is correct and display that.