Make "conditions" a dictionary.
This was causing some inefficiencies when post-processing the user
input.  Number one reason for this: we don't care what order the
conditions are in, and we have to do work on some of them.  This is
better than scanning the whole list of conditions for the "type" and
"color" conditions.  Now we can just get them by name.