Clean up snow-mana display.
For some reason, Gatherer's text mode displays a little "i" in the
snow mana symbols.  We were copying that verbatim for a while, but it
makes the whole thing harder to read and I don't even know why it's
there.  Now snow mana is cleaner and easier to read as just {S}.